keeping pybel Molecule references around and improving encapsulation
I was told through an issue I raised on Open Babel's GitHub that
references to pybel Molecules must be maintained in order for their Atom
instances to be valid so the code has been updated to reflect that.
Prior to that I was getting memory errors when trying to access pybel
Atom attributes. MLAtom now keeps a reference of its parent pybel
molecule. Unrelated, also added methods to report length of feature
vectors so that aggregating classes are not tied to each others'
implementations.